pyproject.toml
[project]
name = "aws-analytics-scripts"
version = "0.1.0"
description = "AWS Analytics Scripts for MSK, OpenSearch, Kinesis, and more"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-msk-iam-sasl-signer-python>=1.0.2",
    "boto3>=1.39.3",
    "faker>=37.5.3",
    "gremlinpython>=3.7.3",
    "kafka-python>=2.2.15",
    "numpy>=1.24.0",
    "opensearch-py>=3.0.0",
    "packaging>=25.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
    "requests-aws4auth>=1.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aws_analytics"]

[project.scripts]
msk-load-test = "scripts.msk_load_test:main"
http-request = "scripts.http_request:main"